Implement remaining 13 fixes from blind spots roster
- Fix #2/#6: Replace HttpWebRequest with shared static HttpClient; log and throw on API errors instead of silently returning defaults
- Fix #5/#9: Add ReferenceDataCache using MemoryDataCacheManager with 5-min TTL for all lookup data across controllers
- Fix #8: Parallelize Monthly View N+1 employee loop with Parallel.ForEach
- Fix #11: Add global ValidateModelAttribute action filter to reject invalid ModelState on API
- Fix #20: Change repository Unity registration from singleton to per-request (HierarchicalLifetimeManager)
- Fix #14: Align TimeEntry.WorkHour from double to decimal to match API-side type
- Fix #18: Migrate AutoMapper from deprecated static Mapper.Initialize() to instance-based MapperConfiguration
- Fix #19: Add X-Correlation-ID header to all API calls; add GlobalExceptionFilterAttribute for unhandled exceptions
- Fix #7/#15: Add Unity DI container to web frontend; all controllers now use constructor injection
- Fix #12: Add ApiKeyAuthHandler for API key authentication (backward-compatible when unconfigured)

+++ Main/Web/Source/SBS.IT.Utilities.Web.TimeTrackerWeb/Models/TimeSheetModel.cs
@@ -28,7 +28,7 @@ namespace SBS.IT.Utilities.Web.TimeTrackerWeb.Models
         public int? ProjectItemId { get; set; }
         public int? ProjectId { get; set; }
         public DateTime Date { get; set; }
-        public double WorkHour { get; set; }
+        public decimal WorkHour { get; set; }
         public string Comments { get; set; }
         public string WeekDays { get; set; }
         public int UserId { get; set; }
